+++ frontend/src/pages/main/_components/index.ts
@@ -2,3 +2,4 @@ export * from './UrlInputForm'
 export * from './VideoCard'
 export * from './VideoRecommendation'
 export * from './ErrorToast'
+export * from './UrlInputModal'
